Make wallet state management & hashing scalable
The commit changes the way wallets are inserted, updated, and hashed, but does not change
the way they are stored on disk or transferred to peers.

The commit includes a hard fork (2.2) required for switching to the new recursive hashing
method.

Wallets are now stored in a tree, closely resembling Patricia Merkle trees.
Each node in the tree stores the hash of its subtree, what allows to efficiently
recompute the hash of the whole collection - it scales with the number of transactions in
the block, not with the total number of wallets in the weave. Such a tree has an advantage over
the binary tree because there is a hard cap on the number of operations which may be
required for recomputing the root hash after a single insert - 32 (key length) * 2 ^ 8.
A binary tree, on the other hand, requires periodic rebalancing, otherwise the complexity
of operations may approach O(number of wallets).